agent/pyproject.toml
[project]
name = "logfire-docs-chatbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "duckdb>=1.4.4",
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "logfire[fastapi,starlette]>=4.9.0",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "pydantic-ai-slim[anthropic,cli,google,openai]>=1.14.0",
    "pydantic-settings>=2.13.1",
    "redis==6.2.0",
    "redislite>=6.2.912183",
    "sqlmodel>=0.0.37",
    "taskiq-fastapi>=0.4.0",
    "taskiq-redis>=1.1.2",
    "uvicorn>=0.37.0",
]

[tool.ruff.format]
# don't format python in docstrings, pytest-examples takes care of it
docstring-code-format = false
quote-style = "single"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]
ignore = [
    "B008", # FastAPI dependency patterns commonly use call defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ty.src]
include = [
    "chatbot",
]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.terminal]
output-format = "concise"

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "ty>=0.0.18",
]
